+++ ui/src/main/java/com/dat3m/ui/options/utils/Task.java
@@ -1,4 +1,4 @@
-package com.dat3m.ui.utils;
+package com.dat3m.ui.options.utils;
 
 public enum Task {
